Section 22.378(Working with Databases: JPA Challenge): Bonus challenge progress so far 2.
not much changed; there are '18' songs matching for '%Soul%'. noticed that JPQL would output '18', when there are only 12 artists. It was a SELECT * FROM artists, so it would duplicate some artists.

still cannot resolve CriteriaBuilder to select for only Artists of Songs where an Album had a Song matching '%Soul%'